segregação do backend/frontend; primeiro componente de navegação e de página

+++ .docker/entrypoint.sh
@@ -2,10 +2,23 @@
 
 #On error no such file entrypoint.sh, execute in terminal - dos2unix .docker\entrypoint.sh
 
+
+npm config set cache /var/www/.npm-cache --global
+cd /var/www/frontend && npm install && cd ..
+
+
+cd backend
+
 chown -R www-data:www-data .
 
-cp .env.example .env
-cp .env.testing.example .env.testing
+if [ ! -f ".env"]; then
+  cp .env.example .env 
+fi
+
+if [ ! -f ".env.testing"]; then
+  cp .env.testing.example .env.testing
+fi
+
 composer install
 php artisan key:generate
 php artisan migrate --seed
